Complete new background daemon handling
Using the unicorn-worker-killer gem we respawn the worker
process every 150-200 requests. Respawning this worker causes
the background daemons to restart too and refresh the database
connections. A request is done every cycle of a background daemon.

This ensures the stability of background daemons and avoid them
to hang. Respawning the web worker processes does not affect to
deploying functionality.