Rewrite the way ASYD has to interact with the clients by SSH
Before it opened a new SSH connection for every command, now it opens a global SSH connection and stores
it on the host class while the deploy it's using it, thus avoiding connection spam. This should represent
a huge performance improvement on environments with heavy deploys, group of hosts and rate-limited networks.